Swift solution to Day 7
This was quite the workout. At first I thought I could have a base Node class
and inherit from there with two classes: File and Directory. The advantage is
that I then could have a single array of Nodes as children of Directory
(whereas File would not have children at all) on which I could fetch a single
Node::size computed property that would either give me the size of a File, or
calculate the total size of a Directory, recurring down the subtree as
necessary. It's a neat solution but it would lead to some ugly code down the
line, so someone suggested to make Node a protocol which required quite the
rewrite but made it for neater code down the line. Swift really doesn't like
to make basic stuff easy for developers.